TGUI Input Framework

+++ code/modules/tgui_input/list.dm
@@ -6,17 +6,17 @@
  * * user - The user to show the input box to.
  * * message - The content of the input box, shown in the body of the TGUI window.
  * * title - The title of the input box, shown on the top of the TGUI window.
- * * buttons - The options that can be chosen by the user, each string is assigned a button on the UI.
+ * * items - The options that can be chosen by the user, each string is assigned a button on the UI.
  * * default - The option with this value will be selected on first paint of the TGUI window.
  * * timeout - The timeout of the input box, after which the input box will close and qdel itself. Set to zero for no timeout.
  */
-/proc/tgui_input_list(mob/user, message, title, list/buttons, default, timeout = 0)
+/proc/tgui_input_list(mob/user, message, title = "Select", list/items, default, timeout = 0)
 	if (istext(user))
 		stack_trace("tgui_alert() received text for user instead of mob")
 		return
 	if (!user)
 		user = usr
-	if(!length(buttons))
+	if(!length(items))
 		return
 	if (!istype(user))
 		if (istype(user, /client))
@@ -24,42 +24,15 @@
 			user = client.mob
 		else
 			return
-	var/datum/tgui_list_input/input = new(user, message, title, buttons, default, timeout)
+	/// Client does NOT have tgui_input on: Returns regular input
+	if(!usr.client.prefs.tgui_input_mode)
+		return input(user, message, title, default) as null|anything in items
+	var/datum/tgui_list_input/input = new(user, message, title, items, default, timeout)
 	input.tgui_interact(user)
 	input.wait()
 	if (input)
 		. = input.choice
 		qdel(input)
-
-/**
- * Creates an asynchronous TGUI input list window with an associated callback.
- *
- * This proc should be used to create inputs that invoke a callback with the user's chosen option.
- * Arguments:
- * * user - The user to show the input box to.
- * * message - The content of the input box, shown in the body of the TGUI window.
- * * title - The title of the input box, shown on the top of the TGUI window.
- * * buttons - The options that can be chosen by the user, each string is assigned a button on the UI.
- * * default - The option with this value will be selected on first paint of the TGUI window.
- * * callback - The callback to be invoked when a choice is made.
- * * timeout - The timeout of the input box, after which the menu will close and qdel itself. Set to zero for no timeout.
- */
-/proc/tgui_input_list_async(mob/user, message, title, list/buttons, default, datum/callback/callback, timeout = 60 SECONDS)
-	if (istext(user))
-		stack_trace("tgui_alert() received text for user instead of mob")
-		return
-	if (!user)
-		user = usr
-	if(!length(buttons))
-		return
-	if (!istype(user))
-		if (istype(user, /client))
-			var/client/client = user
-			user = client.mob
-		else
-			return
-	var/datum/tgui_list_input/async/input = new(user, message, title, buttons, default, callback, timeout)
-	input.tgui_interact(user)
 
 /**
  * # tgui_list_input
@@ -72,14 +45,14 @@
 	var/title
 	/// The textual body of the TGUI window
 	var/message
-	/// The list of buttons (responses) provided on the TGUI window
-	var/list/buttons
-	/// Buttons (strings specifically) mapped to the actual value (e.g. a mob or a verb)
-	var/list/buttons_map
-	/// Value of the button that should be pre-selected on first paint.
-	var/initial
+	/// The list of items (responses) provided on the TGUI window
+	var/list/items
+	/// Items (strings specifically) mapped to the actual value (e.g. a mob or a verb)
+	var/list/items_map
 	/// The button that the user has pressed, null if no selection has been made
 	var/choice
+	/// The default item to be selected
+	var/default
 	/// The time at which the tgui_list_input was created, for displaying timeout progress.
 	var/start_time
 	/// The lifespan of the tgui_list_input, after which the window will close and delete itself.
@@ -87,29 +60,29 @@
 	/// Boolean field describing if the tgui_list_input was closed by the user.
 	var/closed
 
-/datum/tgui_list_input/New(mob/user, message, title, list/buttons, default, timeout)
+/datum/tgui_list_input/New(mob/user, message, title, list/items, default, timeout)
 	src.title = title
 	src.message = message
-	src.buttons = list()
-	src.buttons_map = list()
-	src.initial = default
-	var/list/repeat_buttons = list()
+	src.items = list()
+	src.items_map = list()
+	src.default = default
+	var/list/repeat_items = list()
 
 	// Gets rid of illegal characters
 	var/static/regex/whitelistedWords = regex(@{"([^\u0020-\u8000]+)"})
 
-	for(var/i in buttons)
+	for(var/i in items)
 		if(isnull(i))
-			stack_trace("Null in a tgui_input_list() buttons")
+			stack_trace("Null in a tgui_input_list() items")
 			continue
 
 		var/string_key = whitelistedWords.Replace("[i]", "")
 
 		//avoids duplicated keys E.g: when areas have the same name
-		string_key = avoid_assoc_duplicate_keys(string_key, repeat_buttons)
+		string_key = avoid_assoc_duplicate_keys(string_key, repeat_items)
 
-		src.buttons += string_key
-		src.buttons_map[string_key] = i
+		src.items += string_key
+		src.items_map[string_key] = i
 
 
 	if (timeout)
@@ -119,7 +92,7 @@
 
 /datum/tgui_list_input/Destroy(force, ...)
 	SStgui.close_uis(src)
-	QDEL_NULL(buttons)
+	QDEL_NULL(items)
 	. = ..()
 
 /**
@@ -133,7 +106,7 @@
 /datum/tgui_list_input/tgui_interact(mob/user, datum/tgui/ui)
 	ui = SStgui.try_update_ui(user, src, ui)
 	if(!ui)
-		ui = new(user, src, "ListInput")
+		ui = new(user, src, "ListInputModal")
 		ui.open()
 
 /datum/tgui_list_input/tgui_close(mob/user)
@@ -144,27 +117,31 @@
 	return GLOB.tgui_always_state
 
 /datum/tgui_list_input/tgui_static_data(mob/user)
-	. = list(
-		"title" = title,
-		"message" = message,
-		"buttons" = buttons,
-		"initial" = initial
-	)
+	var/list/data = list()
+	data["init_value"] = default || items[1]
+	data["items"] = items
+	data["large_buttons"] = usr.client.prefs.tgui_large_buttons
+	data["message"] = message
+	data["large_buttons"] = usr.client.prefs.tgui_swapped_buttons
+	data["title"] = title
+	return data
 
 /datum/tgui_list_input/tgui_data(mob/user)
-	. = list()
+	var/list/data = list()
 	if(timeout)
 		.["timeout"] = clamp((timeout - (world.time - start_time) - 1 SECONDS) / (timeout - 1 SECONDS), 0, 1)
+	return data
 
 /datum/tgui_list_input/tgui_act(action, list/params)
 	. = ..()
 	if (.)
 		return
 	switch(action)
-		if("choose")
-			if (!(params["choice"] in buttons))
+		if("submit")
+			if (!(params["entry"] in items))
 				return
-			set_choice(buttons_map[params["choice"]])
+			set_choice(items_map[params["entry"]])
+			closed = TRUE
 			SStgui.close_uis(src)
 			return TRUE
 		if("cancel")
@@ -176,6 +153,36 @@
 	src.choice = choice
 
 /**
+ * Creates an asynchronous TGUI input list window with an associated callback.
+ *
+ * This proc should be used to create inputs that invoke a callback with the user's chosen option.
+ * Arguments:
+ * * user - The user to show the input box to.
+ * * message - The content of the input box, shown in the body of the TGUI window.
+ * * title - The title of the input box, shown on the top of the TGUI window.
+ * * items - The options that can be chosen by the user, each string is assigned a button on the UI.
+ * * default - The option with this value will be selected on first paint of the TGUI window.
+ * * callback - The callback to be invoked when a choice is made.
+ * * timeout - The timeout of the input box, after which the menu will close and qdel itself. Set to zero for no timeout.
+ */
+/proc/tgui_input_list_async(mob/user, message, title, list/items, default, datum/callback/callback, timeout = 60 SECONDS)
+	if (istext(user))
+		stack_trace("tgui_alert() received text for user instead of mob")
+		return
+	if (!user)
+		user = usr
+	if(!length(items))
+		return
+	if (!istype(user))
+		if (istype(user, /client))
+			var/client/client = user
+			user = client.mob
+		else
+			return
+	var/datum/tgui_list_input/async/input = new(user, message, title, items, default, callback, timeout)
+	input.tgui_interact(user)
+
+/**
  * # async tgui_list_input
  *
  * An asynchronous version of tgui_list_input to be used with callbacks instead of waiting on user responses.
@@ -184,8 +191,8 @@
 	/// The callback to be invoked by the tgui_list_input upon having a choice made.
 	var/datum/callback/callback
 
-/datum/tgui_list_input/async/New(mob/user, message, title, list/buttons, default, callback, timeout)
-	..(user, title, message, buttons, default, timeout)
+/datum/tgui_list_input/async/New(mob/user, message, title, list/items, default, callback, timeout)
+	..(user, title, message, items, default, timeout)
 	src.callback = callback
 
 /datum/tgui_list_input/async/Destroy(force, ...)
